reduce state.go diff
