+++ vms/platformvm/state/state.go
@@ -10,6 +10,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/google/btree"
 	"github.com/prometheus/client_golang/prometheus"
 
 	"go.uber.org/zap"
@@ -1141,6 +1142,95 @@ func (s *state) syncGenesis(genesisBlk block.Block, genesis *genesis.Genesis) er
 	// primary network's validator set is empty before the validator sets are
 	// initialized.
 	return s.write(false /*=updateValidators*/, 0)
+}
+
+// Load pulls data previously stored on disk that is expected to be in memory.
+func (s *state) load(hasSynced bool) error {
+	return utils.Err(
+		s.loadMerkleMetadata(),
+		s.loadCurrentStakers(),
+		s.loadPendingStakers(),
+		s.initValidatorSets(),
+
+		s.logMerkleRoot(!hasSynced), // we already logged if sync has happened
+	)
+}
+
+// Loads the chain time and last accepted block ID from disk
+// and populates them in [ms].
+func (s *state) loadMerkleMetadata() error {
+	// load chain time
+	chainTimeBytes, err := s.merkleDB.Get(merkleChainTimeKey)
+	if err != nil {
+		return err
+	}
+	var chainTime time.Time
+	if err := chainTime.UnmarshalBinary(chainTimeBytes); err != nil {
+		return err
+	}
+	s.latestComittedChainTime = chainTime
+	s.SetTimestamp(chainTime)
+
+	// load last accepted block
+	blkIDBytes, err := s.merkleDB.Get(merkleLastAcceptedBlkIDKey)
+	if err != nil {
+		return err
+	}
+	lastAcceptedBlkID := ids.Empty
+	copy(lastAcceptedBlkID[:], blkIDBytes)
+	s.latestCommittedLastAcceptedBlkID = lastAcceptedBlkID
+	s.SetLastAccepted(lastAcceptedBlkID)
+
+	// We don't need to load supplies. Unlike chain time and last block ID,
+	// which have the persisted* attribute, we signify that a supply hasn't
+	// been modified by making it nil.
+	return nil
+}
+
+// Loads current stakes from disk and populates them in [ms].
+func (s *state) loadCurrentStakers() error {
+	// TODO ABENEGIA: Check missing metadata
+	s.currentStakers = newBaseStakers()
+
+	prefix := make([]byte, len(currentStakersSectionPrefix))
+	copy(prefix, currentStakersSectionPrefix)
+
+	iter := s.merkleDB.NewIteratorWithPrefix(prefix)
+	defer iter.Release()
+	for iter.Next() {
+		data := &stakersData{}
+		if _, err := txs.GenesisCodec.Unmarshal(iter.Value(), data); err != nil {
+			return fmt.Errorf("failed to deserialize current stakers data: %w", err)
+		}
+
+		tx, err := txs.Parse(txs.GenesisCodec, data.TxBytes)
+		if err != nil {
+			return fmt.Errorf("failed to parsing current stakerTx: %w", err)
+		}
+		stakerTx, ok := tx.Unsigned.(txs.Staker)
+		if !ok {
+			return fmt.Errorf("expected tx type txs.Staker but got %T", tx.Unsigned)
+		}
+
+		staker, err := NewCurrentStaker(tx.ID(), stakerTx, data.PotentialReward)
+		if err != nil {
+			return err
+		}
+		if staker.Priority.IsValidator() {
+			// TODO: why not PutValidator/PutDelegator??
+			validator := s.currentStakers.getOrCreateValidator(staker.SubnetID, staker.NodeID)
+			validator.validator = staker
+			s.currentStakers.stakers.ReplaceOrInsert(staker)
+		} else {
+			validator := s.currentStakers.getOrCreateValidator(staker.SubnetID, staker.NodeID)
+			if validator.delegators == nil {
+				validator.delegators = btree.NewG(defaultTreeDegree, (*Staker).Less)
+			}
+			validator.delegators.ReplaceOrInsert(staker)
+			s.currentStakers.stakers.ReplaceOrInsert(staker)
+		}
+	}
+	return iter.Error()
 }
 
 func (s *state) GetCurrentDelegatorIterator(subnetID ids.ID, nodeID ids.NodeID) (StakerIterator, error) {
